Start of Docker rewrite, add mod log monitoring
This begins a large-scale shift to using containers with Docker and docker-compose.

It also moves to using Postgres with alembic for database migrations
and a lightweight home-made ORM instead of sqlalchemy models.

The first use of this new infrastructure is mod activity through the mod_log script.

Config is still a mess through the transition. Eventually .env files will be predominant
but for now config.py is still widely used.